Deleted unused test code.(#11)

+++ iconservice/iconscore/icon_score_installer.py
@@ -13,18 +13,21 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+import logging
 import os
 import zipfile
 import shutil
+import pip
 
-from icon.iconservice.base.address import Address
+from ..base.address import Address
+from ..base.exception import ScoreInstallException, ScoreInstallExtractException
+from ..base.exception import ScoreInstallWriteZipfileException
 
-CONST_FILE_NOT_FOUND_ERROR_CODE = 912
-CONST_IS_A_DIRECTORY_ERROR_CODE = 913
-CONST_PERMISSION_ERROR_CODE = 914
-CONST_NOT_A_DIRECTORY_ERROR_CODE = 915
-CONST_BAD_ZIP_FILE_ERROR_CODE = 916
-CONST_FILE_EXISTS_ERROR__CODE = 917
+CONST_SCORE_EXISTS_ERROR_CODE = 918
+CONST_WRITE_ZIPFILE_ERROR_CODE = 919
+CONST_EXTRACT_FILES_ERROR_CODE = 920
+CONST_SUCCESS_INSTALL_SCORE = 910
+CONST_PERMISSION_ERROR_CODE = 911
 
 
 class IconScoreInstaller(object):
@@ -34,20 +37,55 @@ class IconScoreInstaller(object):
     def __init__(self, icon_score_root_path: str) -> None:
         self.icon_score_root_path = icon_score_root_path
 
-    def install(self, address: Address, data: bytes) -> None:
+    def install(self, address: Address, data: bytes, block_height: int, transaction_index: int) -> int:
         """Install score.
 
         :param address: contract address
         :param data: The byte value of the zip file.
+        :param block_height:
+        :param transaction_index:
         :return:
         """
-        install_path = os.path.join(self.icon_score_root_path, str(address))
-        zip_path = IconScoreInstaller.write_zipfile_with_bytes(install_path, data)
-        zip_root_directory_name = IconScoreInstaller.extract_files(self.icon_score_root_path, zip_path)
+        str_block_height = str(block_height)
+        str_transaction_index = str(transaction_index)
+        str_address = str(address)
+        score_id = str_block_height + "_" + str_transaction_index
+        score_root_path = os.path.join(self.icon_score_root_path, str_address)
+        install_path = os.path.join(score_root_path, score_id)
+        zip_path = install_path + ".zip"
+        try:
+            if not os.path.exists(score_root_path):
+                os.makedirs(score_root_path)
+            if os.path.exists(install_path):
+                raise ScoreInstallException(f"You already have a SCORE of this version.")
+            zip_path = IconScoreInstaller.write_zipfile_with_bytes(zip_path, data)
+            zip_root_directory_name = IconScoreInstaller.extract_files(self.icon_score_root_path, zip_path)
 
-        IconScoreInstaller.remove_exists_archive(zip_path)
-        shutil.move(os.path.join(self.icon_score_root_path, zip_root_directory_name),
-                    os.path.join(self.icon_score_root_path, str(address)))
+            IconScoreInstaller.remove_exists_archive(zip_path)
+            shutil.move(os.path.join(self.icon_score_root_path, zip_root_directory_name),
+                        install_path)
+            # IconScoreInstaller.install_requirements(install_path)
+            return CONST_SUCCESS_INSTALL_SCORE
+        except ScoreInstallException as e:
+            logging.debug(e.message)
+            return CONST_SCORE_EXISTS_ERROR_CODE
+        except ScoreInstallWriteZipfileException:
+            return CONST_WRITE_ZIPFILE_ERROR_CODE
+        except ScoreInstallExtractException:
+            return CONST_EXTRACT_FILES_ERROR_CODE
+        except PermissionError as pe:
+            logging.debug(pe)
+            return CONST_PERMISSION_ERROR_CODE
+
+    @staticmethod
+    def install_requirements(install_path: str):
+        for current_dir, dirs, file_list in os.walk(install_path):
+            if 'requirements.txt' in file_list:
+                with open(os.path.join(install_path, 'requirements.txt'), 'r') as requires:
+                    requirements_list_str = requires.read()
+                for package in requirements_list_str.split("\n"):
+                    if bool(package):
+                        pip.main(['install', package])
 
     @staticmethod
     def extract_files(install_path: str, archive_path: str) -> str:
@@ -59,36 +97,43 @@ class IconScoreInstaller(object):
         Will return root directory of SCORE.
         """
         try:
-            zip_file = zipfile.ZipFile(archive_path, 'r')
-            file_name_list = zip_file.namelist()
-            file_name_prefix_list = [prefix.split('/')[0] for prefix in file_name_list
-                                     if not prefix.startswith('__MACOSX')]
-            is_zipfile_covered = all(a == file_name_prefix_list[0] for a in file_name_prefix_list)
-            zip_root_name = file_name_list[0].split("/")[0]
+            if zipfile.is_zipfile(archive_path) is False:
+                IconScoreInstaller.remove_exists_archive(archive_path)
+                raise ScoreInstallExtractException("The data you sent is not a zip file.")
+            else:
+                with zipfile.ZipFile(archive_path, 'r') as z:
+                    file_name_list = z.namelist()
+                    file_name_prefix_list = [prefix.split('/')[0] for prefix in file_name_list
+                                             if not prefix.startswith('__MACOSX')]
+                    is_zipfile_covered = all(a == file_name_prefix_list[0] for a in file_name_prefix_list)
+                    zip_root_name = file_name_list[0].split("/")[0]
 
-            if is_zipfile_covered:
-                IconScoreInstaller._extract_files(file_name_list, install_path, zip_file)
-                return zip_root_name
-
-            IconScoreInstaller._extract_files(file_name_list, install_path + '/temp', zip_file)
-            return 'temp'
+                    if is_zipfile_covered:
+                        IconScoreInstaller._extract_files(file_name_list, install_path, z)
+                        return zip_root_name
+                    IconScoreInstaller._extract_files(file_name_list, install_path + '/temp', z)
+                    return 'temp'
 
         except FileNotFoundError:
-            print(f"{archive_path} not found. check file path.")
-            return CONST_FILE_NOT_FOUND_ERROR_CODE
+            logging.debug(f"{archive_path} not found. check file path.")
+            raise ScoreInstallExtractException(f"{archive_path} not found. check file path.")
         except IsADirectoryError:
-            print(f"{archive_path} is a directory. check file path.")
-            return CONST_IS_A_DIRECTORY_ERROR_CODE
+            logging.debug(f"{archive_path} is a directory. check file path.")
+            raise ScoreInstallExtractException(f"{archive_path} is a directory. check file path.")
         except PermissionError:
-            return CONST_PERMISSION_ERROR_CODE
+            raise ScoreInstallExtractException("Permission error")
         except NotADirectoryError:
-            print(f"{install_path} is not a directory")
-            return CONST_NOT_A_DIRECTORY_ERROR_CODE
+            logging.debug(f"{install_path} is not a directory")
+            raise ScoreInstallExtractException(f"{install_path} is not a directory")
         except zipfile.BadZipFile:
-            print(f"{archive_path}, bad zip file.")
-            return CONST_BAD_ZIP_FILE_ERROR_CODE
-        finally:
-            zip_file.close()
+            logging.debug(f"{archive_path} is bad zip file.")
+            raise ScoreInstallExtractException(f"{archive_path} is bad zip file.")
+        except zipfile.LargeZipFile:
+            logging.debug(f"{archive_path} is too Large Zipfile.")
+            raise ScoreInstallExtractException(f"{archive_path} is too Large file.")
+        except Exception as e:
+            logging.debug(e)
+            raise ScoreInstallExtractException("Exception occurred while extracting zip file.")
 
     @staticmethod
     def _extract_files(file_name_list: str, install_path: str, zip_file: object) -> None:
@@ -124,22 +169,23 @@ class IconScoreInstaller(object):
         :return:
         """
         if os.path.isfile(archive_path):
-            print(f"{archive_path} file exists.")
-            return CONST_FILE_EXISTS_ERROR__CODE
+            logging.debug(f"{archive_path} file exists.")
+            raise ScoreInstallWriteZipfileException(f"{archive_path} file exists.")
         try:
             with open(archive_path, 'wb') as f:
                 f.write(byte_data)
             return archive_path
         except IsADirectoryError:
-            print(f"{archive_path} is a directory. check file path.")
-            return CONST_IS_A_DIRECTORY_ERROR_CODE
+            logging.debug(f"{archive_path} is a directory. check file path.")
+            raise ScoreInstallWriteZipfileException(f"{archive_path} is a directory. check file path.")
         except PermissionError:
-            print(f"permission error")
-            return CONST_PERMISSION_ERROR_CODE
+            logging.debug(f"permission error")
+            print(archive_path)
+            return ScoreInstallWriteZipfileException(f"permission error")
 
     # This method will be removed. written for test.
     @staticmethod
     def read_zipfile_as_byte(archive_path: str) -> bytes:
         with open(archive_path, 'rb') as f:
             byte_data = f.read()
-        return byte_data
+            return byte_data
